move and rename popgenmut
