+++ fwdpp/fundamental_types/Makefile.am
@@ -1,3 +1,5 @@
+pkgincludedir=$(prefix)/include/fwdpp/fundamental_types
+
 pkginclude_HEADERS=typedefs.hpp \
 				   haploid_genome.hpp \
 				   mutation_base.hpp
