+++ fwdpp/types/Makefile.am
@@ -1,0 +1,4 @@
+pkgincludedir=$(prefix)/include/fwdpp/types
+
+pkginclude_HEADERS=mutation.hpp
+
